refactor: fix typo in resources
add fileStorage to prettier ignore
