+++ visualization/app/codeCharta/resources/fixed-folders/fixed-folders-example.ts
