+++ visualization/app/codeCharta/ui/dialog/dialog.globalSettings.e2e.ts
@@ -16,7 +16,7 @@ describe("DialogGlobalSettings", () => {
 	})
 
 	async function setupTest() {
-		await fileChooser.openFiles(["./app/codeCharta/ressources/sample1_with_different_edges.cc.json"])
+		await fileChooser.openFiles(["./app/codeCharta/resources/sample1_with_different_edges.cc.json"])
 		await globalSettingsPageObject.openGlobalSettings()
 	}
